+++ frontend/src/pages/DetailedReviewPage/components/index.tsx
@@ -1,3 +1,4 @@
 export { default as KeywordSection } from './KeywordSection';
 export { default as ReviewDescription } from './ReviewDescription';
 export { default as ReviewSection } from './ReviewSection';
+export { default as DetailedReviewPageContents } from './DetailedReviewPageContents';
